Extract Mongoid::Base from Documents so that we get InvalidFieldErrors when we try to create fields conflicting with those methods
