+++ lib/mongoid.rb
@@ -37,6 +37,7 @@ require "mongoid/threaded"
 require "mongoid/relations"
 require "mongoid/atomic"
 require "mongoid/attributes"
+require "mongoid/base"
 require "mongoid/callbacks"
 require "mongoid/collection"
 require "mongoid/collections"
